Removed RestInterfaceTestBase, because it's only used in one place.
